попытки задетектить коллизию

+++ Assets/Scripts/Systems/ShootingSystem.cs
@@ -1,6 +1,8 @@
 ﻿using Data;
 using Data.Events;
 using Unity.Entities;
+using Unity.Mathematics;
+using Unity.Transforms;
 
 namespace Systems
 {
@@ -8,12 +10,29 @@ namespace Systems
     {
         protected override void OnUpdate()
         {
-            Entities.ForEach((Entity entity, in FireRequestEvent fireRequestEvent, in WeaponData shootData) =>
+            Entities.WithAll<FireRequestEvent>().ForEach((Entity entity, ref ShootData shootData, in LocalToWorld localToWorld, in WeaponData weaponData, in TargetData targetData) =>
             {
-                var firePointEntity = shootData.FirePoint;
-                EntityManager.AddComponentData(firePointEntity, new BulletSpawnEvent());
+                if (targetData.Value == Entity.Null)
+                    return;
+                
+                var firePointEntity = weaponData.FirePoint;
+                float3 firePointPosition = EntityManager.GetComponentData<LocalToWorld>(firePointEntity).Position;
+                float3 direction = GetFireDirection(targetData, localToWorld);
+                shootData.FirePosition = firePointPosition;
+                shootData.Direction = direction;
+
+                EntityManager.AddComponentData(entity, new BulletSpawnEvent());
                 EntityManager.RemoveComponent(entity, typeof(FireRequestEvent));
             }).WithStructuralChanges().Run();
+        }
+
+        private float3 GetFireDirection(TargetData targetData, LocalToWorld localToWorld)
+        {
+            float3 direction;
+            LocalToWorld targetLocalToWorld = EntityManager.GetComponentData<LocalToWorld>(targetData.Value);
+            direction = targetLocalToWorld.Position - localToWorld.Position;
+            direction = math.normalize(direction);
+            return direction;
         }
     }
 }
